Updated data type for repositories

+++ FRS/src/main/java/com/group/repository/Flight_ScheduleRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.group.model.Flight_Schedule;
 
-public interface Flight_ScheduleRepository extends CrudRepository<Flight_Schedule, Long> {
+public interface Flight_ScheduleRepository extends CrudRepository<Flight_Schedule, Integer> {
