+++ FRS/src/main/java/com/group/repository/Passenger_DetailRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.group.model.Passenger_Detail;
 
-public interface Passenger_DetailRepository  extends CrudRepository<Passenger_Detail, Long> {
+public interface Passenger_DetailRepository  extends CrudRepository<Passenger_Detail, Integer> {
